2025.12.12_관리자/과정템플릿/과정기수/수강신청관리_유혜림

+++ src/main/java/com/gocle/yangju/forest/adm/company/service/impl/CompanyManageServiceImpl.java
@@ -49,6 +49,16 @@ public class CompanyManageServiceImpl extends EgovAbstractServiceImpl implements
 		    companyManageMapper.insertCompanyLocation(companyManageVo);
 		}
 		
+		String[] cateCds = companyManageVo.getCateCd().split(",");
+		for (int i = 0; i < cateCds.length; i++) {
+			String cc = cateCds[i].trim(); // 공백 제거
+			String sgrCd = cc.split("_")[0];
+			String cateCd = cc.split("_")[1];
+			companyManageVo.setSgrCd(sgrCd);
+			companyManageVo.setCateCd(cateCd);
+			companyManageMapper.insertCompanyMapping(companyManageVo);
+		}
+		
 		return data;
 	}
 
@@ -63,19 +73,34 @@ public class CompanyManageServiceImpl extends EgovAbstractServiceImpl implements
 		
 		int data = companyManageMapper.updateBoardArticle(companyManageVo);
 		
-		companyManageMapper.deleteCompanyLocation(companyManageVo);
+		if(companyManageVo.getLocation() != null && companyManageVo.getLocation() != "") {
+			companyManageMapper.deleteCompanyLocation(companyManageVo);
+			
+			String[] locations = companyManageVo.getLocation().split(",");
+			
+			for (int i = 0; i < locations.length; i++) {
+	
+			    String loc = locations[i].trim(); // 공백 제거
+			    if (loc.isEmpty()) continue;      // 빈 값은 건너뛰기
+	
+			    companyManageVo.setSeqNo(i + 1);  // i=0 → seq 1
+			    companyManageVo.setLocation(loc); // 각 location 세팅
+	
+			    companyManageMapper.insertCompanyLocation(companyManageVo);
+			}
+		}
 		
-		String[] locations = companyManageVo.getLocation().split(",");
-		
-		for (int i = 0; i < locations.length; i++) {
-
-		    String loc = locations[i].trim(); // 공백 제거
-		    if (loc.isEmpty()) continue;      // 빈 값은 건너뛰기
-
-		    companyManageVo.setSeqNo(i + 1);  // i=0 → seq 1
-		    companyManageVo.setLocation(loc); // 각 location 세팅
-
-		    companyManageMapper.insertCompanyLocation(companyManageVo);
+		if(companyManageVo.getCateCd() != null && companyManageVo.getCateCd() != "") {
+			companyManageMapper.deleteCompanyMapping(companyManageVo);
+			String[] cateCds = companyManageVo.getCateCd().split(",");
+			for (int i = 0; i < cateCds.length; i++) {
+				String cc = cateCds[i].trim(); // 공백 제거
+				String sgrCd = cc.split("_")[0];
+				String cateCd = cc.split("_")[1];
+				companyManageVo.setSgrCd(sgrCd);
+				companyManageVo.setCateCd(cateCd);
+				companyManageMapper.insertCompanyMapping(companyManageVo);
+			}
 		}
 		
 		return data;
@@ -87,6 +112,7 @@ public class CompanyManageServiceImpl extends EgovAbstractServiceImpl implements
 		int data = companyManageMapper.deleteBoardArticle(companyManageVo);
 		
 		companyManageMapper.deleteCompanyLocation(companyManageVo);
+		companyManageMapper.deleteCompanyMapping(companyManageVo);
 		
 		return data;
 	}
@@ -97,4 +123,9 @@ public class CompanyManageServiceImpl extends EgovAbstractServiceImpl implements
 		return companyManageMapper.getLocation(companyManageVo);
 	}
 	
+	@Override
+	public List<CompanyManageVo> getMapping(CompanyManageVo companyManageVo) throws Exception {
+		
+		return companyManageMapper.getMapping(companyManageVo);
+	}
 }
